feat: make entities unique and create them via factories
Each entity has an associated factory that lives in the context. The factory stores a cache of
entities of its type by UUID, so that only one copy of each entity can ever exist. This means we can
update them when some procedure makes changes to it, or based on events, and the client will always
be aware of the changes