feat: allow user to configure transaction speed
The polymath SDK constructor now receives a `speed` parameter that allows setting the default gas
price for transactions. The bigger the gas price, the faster transactions get mined. The possible
settings come in a `TransactionSpeed` enum and are `Slow`, `Medium` and  `Fast`, which correspond to
`safeLow`, `average` and `fastest` in https://ethgasstation.info. For testnets, the gas values are
hardcoded and range from 1 to 5 gwei

close #66